tcetest_bypass.sh: Add error message when itrace not correct

+++ testsuite/systemtest_long/procgen/RISCV/tcetest_bypass.sh
@@ -40,9 +40,11 @@ DIFF_FILE=proge-output/bypass_test.diff
 [ ! -s" $RTL_REG_TRACE" ]     && { echo "FAILURE: RTL reg trace is empty"; exit 1; }
 
 diff ${RTL_ITRACE} ${GOLDEN_ITRACE} > $DIFF_FILE || exit 1
+[ ! -s ${DIFF_FILE} ] && { echo "FAILURE: RTL itrace and golden itrace differ"; exit 1; }
 diff ${RTL_REG_TRACE} ${GOLDEN_REG_TRACE} >> $DIFF_FILE || exit 1
+[ ! -s ${DIFF_FILE} ] && { echo "FAILURE: RTL reg trace and golden reg trace differ"; exit 1; }
 if [ -s $DIFF_FILE ]; then
-    echo BYPASS TEST FAILED
+    echo "BYPASS TEST FAILED"
     exit 1;
 fi
 rm -rf proge-output
